# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: ATANDA Command Centre + HARNESS API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /exemplars/from-artifact:
    post:
      operationId: addArtifactToExemplarLibrary
      tags: [exemplars]
      summary: Copy a completed SPC, MA, MPDD, or PDD artifact into the open library
      requestBody:
        required: true
        content:
          application/json:
            schema:
              {
                $ref: "#/components/schemas/AddArtifactToExemplarLibraryRequest",
              }
      responses:
        "201":
          description: Exemplar added
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ExemplarSummary" }
        "400":
          description: Unsupported artifact type
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          description: Artifact already exists in the library
  /exemplars/{id}:
    get:
      operationId: getExemplar
      tags: [exemplars]
      summary: Fetch a single exemplar including its full markdown body
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Exemplar" }
        "404":
          $ref: "#/components/responses/NotFound"
...
components:
...
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: atanda_staff
  responses:
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
...
    NotFound:
      description: Not found
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
...
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error: { type: string }
        detail: { type: ["string", "null"] }
...
    ExemplarSummary:
      type: object
      required:
        [
          id,
          title,
          tagline,
          jcse,
          certClass,
          source,
          kind,
          sku,
          disc,
          marketplace,
          spcPlayer,
        ]
      properties:
        id: { type: string }
        title: { type: string }
        tagline: { type: string }
        jcse: { type: ["integer", "null"] }
        certClass: { type: ["string", "null"] }
        sku: { type: string }
        source:
          type: string
          enum: [canonical, hand_authored, generated]
        kind:
          type: string
          enum: [SPC, MA, MPDD, PDD]
        marketplace:
          type: string
          enum: [curated, open]
        createdAt: { type: string, format: date-time }
        disc:
          type: ["string", "null"]
          description: "DISC personality profile (Dominance / Influence / Steadiness / Conscientiousness). Every SPC has a unique DISC fingerprint; PDDs are null."
        spcPlayer:
          oneOf:
            - $ref: "#/components/schemas/SpcPlayerPublication"
            - type: "null"
    Exemplar:
      allOf:
        - $ref: "#/components/schemas/ExemplarSummary"
        - type: object
          required: [body]
          properties:
            body: { type: string }
...
    AddArtifactToExemplarLibraryRequest:
      type: object
      required: [artifactId]
      properties:
        artifactId: { type: string, format: uuid }
    SpcQualityScores:
      type: object
      required: [clarity, truthfulness, detectability]
      additionalProperties: false
      description: Three independent advisory axes. Null means the check has not run. No composite score is permitted.
      properties:
        clarity: { type: ["number", "null"] }
        truthfulness: { type: ["number", "null"] }
        detectability: { type: ["number", "null"] }
    SpcEnvironmentNote:
      type: object
      required: [environment, status, note]
      properties:
        environment: { type: string }
        status:
          type: string
          enum: [sketch, verified_integration]
        note: { type: string }
    SpcCheatSheet:
      type: object
      required:
        [
          mission,
          skillSet,
          useCases,
          productionProcess,
          thirdPartyDefinitions,
          environmentNotes,
        ]
      properties:
        mission: { type: string }
        skillSet:
          type: array
          items: { type: string }
        useCases:
          type: array
          items: { type: string }
        productionProcess:
          type: array
          items: { type: string }
        thirdPartyDefinitions:
          type: ["array", "null"]
          items: { type: string }
        environmentNotes:
          type: array
          items: { $ref: "#/components/schemas/SpcEnvironmentNote" }
    SpcPlayerPublication:
      type: object
      required:
        [
          registryVersion,
          publicationStatus,
          cheatSheetPublished,
          publishedBy,
          publishedAt,
          cheatSheet,
          qualityScores,
        ]
      properties:
        registryVersion: { type: string }
        publicationStatus:
          type: string
          enum: [legacy, draft, published]
        cheatSheetPublished: { type: boolean }
        publishedBy:
          type: ["string", "null"]
          enum: [sphinx_engine, null]
        publishedAt: { type: ["string", "null"], format: date-time }
        cheatSheet:
          oneOf:
            - $ref: "#/components/schemas/SpcCheatSheet"
            - type: "null"
        qualityScores: { $ref: "#/components/schemas/SpcQualityScores" }
...